feat: update manga react

+++ source/manga-backend/history_service/src/main/java/com/raindrop/history_service/dto/response/ReadingHistoryResponse.java
@@ -16,7 +16,7 @@ public class ReadingHistoryResponse {
     String mangaId;
     String chapterId;
     String mangaTitle;
-    Integer chapterNumber;
+    double chapterNumber;
     LocalDateTime createdAt;
     LocalDateTime updatedAt;
